M_10_Files/homework/pyproject.toml
[tool.poetry]
name = "shopapp"
version = "0.1.0"
description = ""
authors = ["ramazan_aisaev <aisaev-68@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.6"
python-decouple = "^3.7"
requests = "^2.28.2"
beautifulsoup4 = "^4.11.2"
pillow = "^9.4.0"
mimesis = "^7.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
